iReport+Parse
Finished connection between iReporter and Parse, and user interface can
work now. The problem is that despatch_once didn’t work right now. And,
I need to fix user register and login. Then, the interface of the
commandRequest which will be used to send request to server and get
data as well.